fix filedrank bug. make rank type double.

+++ app-parent/sotouapp/src/main/java/cn/sotou/web/biz/query/criteria/FieldRank.java
@@ -1,22 +1,22 @@
 package cn.sotou.web.biz.query.criteria;
 
 public class FieldRank {
-	private int min;
-	private int max;
+	private double min;
+	private double max;
 
-	public int getMin() {
+	public double getMin() {
 		return min;
 	}
 
-	public void setMin(int min) {
+	public void setMin(double min) {
 		this.min = min;
 	}
 
-	public int getMax() {
+	public double getMax() {
 		return max;
 	}
 
-	public void setMax(int max) {
+	public void setMax(double max) {
 		this.max = max;
 	}
 }
